requirements.txt
tqdm
dill
termcolor
numpy>=1.21
matplotlib>=3.5
scipy>=1.7
scikit_learn>=1.0
torch>=1.10
av>=9.2
torchvision
torchaudio
pytorch_lightning>=1.5
omegaconf
hydra-core==1.2
tensorboard
GPUtil
torch-fidelity
git+https://github.com/domkirke/lardon